Change structure to get around jest limitation
