+++ test/regression/validateExamplesRegression_latestOnly.test.ts
@@ -1,0 +1,6 @@
+import { latestSpecsOnly } from "./specsFilter"
+import { validateExamplesRegressionTest } from "./validateExamplesRegressionTest"
+
+describe("validateExamples should not regress for file", () => {
+  test.each(latestSpecsOnly)("'%s'", validateExamplesRegressionTest, 999999)
+})
