+++ test/regression/validateExamplesRegression_nonLatest.test.ts
@@ -1,0 +1,6 @@
+import { allSpecs } from "./specsFilter"
+import { validateExamplesRegressionTest } from "./validateExamplesRegressionTest"
+
+describe("validateExamples should not regress for file", () => {
+  test.each(allSpecs)("'%s'", validateExamplesRegressionTest, 999999)
+})
